Use fastlog in Histogram::Record
Summary:
Use fastlog from https://code.google.com/p/fastapprox/ instead of log.

According to simple (and quite inaccurate) benchmark, fastlog is three times faster

Benchmark

```
#include <stdint.h>
#include <time.h>
#include <cmath>
#include <iostream>

void print(time_t before) {
  std::cout << time(NULL) - before << std::endl;
}

inline float fastlog2(float x) {
  union { float f; uint32_t i; } vx = { x };
  float y = static_cast<float>(vx.i);
  y *= 1.1920928955078125e-7f;
  return y - 126.94269504f;
}

inline float fastlog(float x) {
  union { float f; uint32_t i; } vx = { x };
  float y = static_cast<float>(vx.i);
  y *= 8.2629582881927490e-8f;
  return y - 87.989971088f;
}

int main() {
  float n = 100000000;
  float sum = 0;
  time_t before = time(NULL);
  for (size_t i = 1; i < n; i+=1) {
    sum += log(static_cast<float>(i));
  }
  print(before);
  std::cout << sum << std::endl;

  sum = 0;
  before = time(NULL);
  for (size_t i = 1; i < n; i+=1) {
    sum += fastlog(static_cast<float>(i));
  }
  print(before);
  std::cout << sum << std::endl;
}
```

Output

```
5
5.36871e+08
2
5.36871e+08
```

Test Plan: make fastlog_test && ./fastlog_test

Reviewers: pja

Reviewed By: pja

Subscribers: dhruba, stupaq-fb, rocketspeed

Differential Revision: https://reviews.facebook.net/D50967

+++ src/util/buffered_storage.cc
@@ -181,11 +181,11 @@ class BufferedLogStorageWorker {
       std::string pilot_prefix = "pilot.buffered_storage.";
       batch_entries = all.AddHistogram(pilot_prefix + "batch_entries",
                                        0.0,
-                                       static_cast<double>(max_entries),
+                                       static_cast<float>(max_entries),
                                        1.0);
       batch_bytes = all.AddHistogram(pilot_prefix + "batch_bytes",
                                      0.0,
-                                     static_cast<double>(max_bytes),
+                                     static_cast<float>(max_bytes),
                                      1.0);
       batch_latency_us = all.AddLatency(pilot_prefix + "batch_latency_us");
     }
